requirements.txt
torch
numpy
pandas
scikit-learn
tqdm
transformers
fastapi
uvicorn
python-dotenv
langchain-openai
deepagents
pydantic
opencv-python
pillow
librosa
